# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "chat-Api",
    "contact": "perlaura64@outlook.com",
    "description": "An API that resembles sending WhatsApp messages",
    "version": "1.0"
  }, 
  "paths": {
...
    "/api/v1/users/login": {
      "post": {
        "tags": [
          "users"
        ],
        "summary": "login user",
        "operationId": "loginuser",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/loginuserrequest"
                  },
                  {
                    "example": {
                      "email": "perlaura64@outlook.com",
                      "password": "Aperlam1964$"
                    }
                  }
                ]
              },
              "example": {
                "email": "perlaura64@outlook.com",
                "password": "Aperlam1964$"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
...
    "/api/v1/messages/conversation/{id}": {
      "post": {
        "tags": [
          "messages"
        ],
        "summary": "create new messages",
        "operationId": "createnewmessages",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      },
      "get": {
        "tags": [
          "messages"
        ],
        "summary": "get conversation messages",
        "operationId": "getconversationmessages",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "style": "simple",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    }
  },
  "components": {
    "schemas": {
...
      "loginuserrequest": {
        "title": "loginuserrequest",
        "required": [
          "email",
          "password"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        },
        "example": {
          "email": "perlaura64@outlook.com",
          "password": "Aperlam1964$"
        }
      }
    }
  },
...
}